% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVennEuler.R
\name{createVennEuler}
\alias{createVennEuler}
\title{A createVennEuler Function}
\usage{
createVennEuler(topTabs, compNames, label = "selected", colFeat = "X",
  colPVal = "P.Value", pval = 0.05, pltR = TRUE, pltPdf = TRUE,
  venn = TRUE, eul = TRUE, csv = TRUE)
}
\arguments{
\item{topTabs}{= list of data frames (which are expected to be in "topTable" shape) to be used as input data}

\item{compNames}{= labels (character vector), corresponding with each input list, to be used in the plots and output table}

\item{label}{= name (character) to be included as global label in all output files created}

\item{colFeat}{= name (character) of the column where to look for the elements to be compared (such as "Name", "Symbol" or "GeneID")}

\item{colPVal}{= name (character) of the column with p-values to be used as filtering criteria, such as "Adj.p.val" or "P.value"}

\item{pval}{= threshold (numeric) value to be applied for filtering by p-value (0.05 means to filter out cases with p-val>0.05)}

\item{pltR}{= TRUE/FALSE, to determine if the plots should be done directly in R}

\item{pltPdf}{= TRUE/FALSE, to determine if the plots should be done in an output PDF file}

\item{venn}{= TRUE/FALSE, to determine whether or not to plot the Venn diagram}

\item{eul}{= TRUE/FALSE, to determine whether or not to plot the Euler diagram}

\item{csv=}{TRUE/FALSE, to determine whether or not to write the table of shared elements as csv output file}
}
\value{
sharedElements (data.frame): table with elements shared by each combination of lists provided
}
\description{
Generates Venn and Euler diagram from lists of elements (top tables) with the corresponding lists of shared and non-shared elements
}
\examples{
## load("topTableList.RData") # this loads the topTableList data object
## load("compNamesList.RData") # this loads the topTableList data object
## compNamesList <- c("A1vsB1","A1vsC1","B1vsC1","A2vsB2","A2vsC2")
sharedElems <- createVennEuler(topTabs = topTableList,
                               compNames = compNamesList,
                               label = "demo",
                               colFeat = "Name",
                               colPVal = "P.Value",
                               pval = 0.01,
                               pltR = TRUE,
                               pltPdf = TRUE,
                               eul = TRUE,
                               venn = TRUE,
                               csv = TRUE)
}
\author{
Miriam Mota <miriam.mota@vhir.org> and Ferran Brianso <ferran.brianso@vhir.org>
}
\references{
Hanbo Chen (2014). VennDiagram: Generate high-resolution Venn and Euler plots. R package version 1.6.9. https://CRAN.R-project.org/package=VennDiagram

Lee Wilkinson (2011). venneuler: Venn and Euler Diagrams. R package version 1.1-0. https://CRAN.R-project.org/package=venneuler
}
\keyword{Euler}
\keyword{EulerDiagram}
\keyword{Venn}
\keyword{VennDiagram}
\keyword{diagram}
\keyword{intersection}
\keyword{lists}
\keyword{sets}
\keyword{union}
\keyword{venneuler}

